requirements.txt
cloudinary==1.25.0
Django==3.1.5
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-rest-auth==0.9.5
django-robots==4.0
django-social-share==2.2.1
django-summernote==0.8.11.6
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
whitenoise==5.2.0

